requirements.txt
mkdocs>=1.5.0
mkdocs-material>=9.4.0
pyyaml>=6.0
markdown>=3.4.0
weasyprint>=60.0
python-docx>=0.8.11
